사용 용량 2차 커밋

+++ src/main/java/outsider/cloudServicePJ/Controller/ContentController/FileUploadController.java
@@ -527,7 +527,7 @@ public class FileUploadController {
 
             FileManageVO fileData = mainMapper.useFileCapacity(vo);
 
-            Integer useStorage = fileData.getUSE_STORAGE();
+            Long useStorage = fileData.getUSE_STORAGE();
 
             result.put("useStorage", useStorage);
         }catch(Exception e){
